// 06-heroesApp/src/app/heroes/interfaces/hero.interface.ts
// 43. Hace la interface de heroes, copiar la respuesta json y usar la extension paste json as codeGenerated by https://quicktype.io

export interface Hero {
  id:               string;
  superhero:        string;
  publisher:        Publisher;
  alter_ego:        string;
  first_appearance: string;
  characters:       string;
}

export enum Publisher {
  DCComics = "DC Comics",
  MarvelComics = "Marvel Comics",
}